requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
cryptography==37.0.4
cycler==0.11.0
Deprecated==1.2.13
discord.py==1.7.3
dj-database-url==1.0.0
Django==4.0.6
django-ckeditor==6.4.2
django-heroku==0.3.1
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-tinymce==3.4.0
fonttools==4.34.4
gunicorn==20.1.0
h11==0.13.0
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.52
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.1
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
spotipy==2.20.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.1.0
yarl==1.7.2
